+++ tests/test-stbt-lint.sh
@@ -111,3 +111,60 @@ test_that_stbt_lint_checks_that_wait_until_argument_is_callable() {
 	EOF
     diff -u lint.expected lint.log
 }
+
+test_that_stbt_lint_checks_frame_parameter_in_frameobject_methods() {
+    cat > test.py <<-EOF
+	from stbt import FrameObject, match, match_text, ocr, is_screen_black
+	
+	def find_boxes(frame=None):
+	    pass
+	
+	class ModalDialog(FrameObject):
+	    @property
+	    def is_visible(self):
+	        return bool(find_boxes())
+	
+	class ErrorDialog(ModalDialog):
+	    @property
+	    def is_visible(self):
+	        return bool(
+	            match("videotestsrc-redblue.png") and
+	            match_text("Error") and
+	            not is_screen_black())
+	
+	    @property
+	    def text(self):
+	        return ocr()
+	
+	class Good(FrameObject):
+	    @property
+	    def is_visible(self):
+	        return bool(find_boxes(self._frame))
+	
+	    @property
+	    def property1(self):
+	        return bool(match("videotestsrc-redblue.png", self._frame))
+	
+	    @property
+	    def property2(self):
+	        return bool(match("videotestsrc-redblue.png", frame=self._frame))
+	
+	    def not_a_property(self):
+	        return bool(match("videotestsrc-redblue.png"))
+	
+	def normal_test():
+	    assert match("videotestsrc-redblue.png")
+	EOF
+    cp "$testdir/videotestsrc-redblue.png" .
+    stbt lint --errors-only test.py > lint.log
+
+    cat > lint.expected <<-'EOF'
+	************* Module test
+	E:  9,20: "find_boxes()" missing "frame" argument (stbt-frame-object-missing-frame)
+	E: 15,12: "match('videotestsrc-redblue.png')" missing "frame" argument (stbt-frame-object-missing-frame)
+	E: 16,12: "match_text('Error')" missing "frame" argument (stbt-frame-object-missing-frame)
+	E: 17,16: "is_screen_black()" missing "frame" argument (stbt-frame-object-missing-frame)
+	E: 21,15: "ocr()" missing "frame" argument (stbt-frame-object-missing-frame)
+	EOF
+    diff -u lint.expected lint.log
+}